THRIFT-2812 Go server adding redundant buffering layer Client: Go Patch: Craig Peterson
This closes #259

It has been pointed out to me that #249 was wrong. It was not needed to allow server sockets to use buffering. The correct way is to pass in a TBufferedTransportFactory to the server. This will create buffered sockets as the processor starts up.

This change creates extra buffering, and is actually a real pain to deactivate. Sorry.